Fixing issues related to the default event for new files
With the introduction of this commit, the correct way to process a single file
is to use framework.ingest_data(None, filename, False).
This trigger an event, by default, specified in the configuration file as
default_ingestion_event

The default is ingest_only which is now a method of the BasePipeline.
It doesn't do anything, but it returns an argument so that the current file being processed is not lost